Implement smooth curves mode
Not Beizer - more like a (more performant and fewer SLoC) cheat.

See https://stackoverflow.com/a/40978275.

This is NOT a general solution for drawing a curve through an ordered
list of arbitrary points, but works for us since we can make the two
assumptions noted in the code comments below.