Feature: Synced lyrics (#54)
* Initial lyrics backend

* Add initial lyrics code

* Implement SQLite side

* Hook lyrics code to cache, player and settings

* First working proof of concept

Lyrics are currently shown in the pane only (queue view), not the bar.

To make way for the thing I've moved the volume knob and output controls
into a GtkPopover. They looked out of place in the pane anyway so this
is also a cleanup in my eyes.

The lyrics widget right now is just a simple ListBox that auto-selects the
currently-sung line. It doesn't even auto-scroll yet. It's just there to
debug the underlying logic.

* Implement autoscroll

* Fix plain lyrics serialisation

* Fully implement widget & controls

* Bump version & tweak focus grabbing

* Println cleanup & update README