v0.94.0-alpha: Local (UNIX) socket support & queue view scroll fix
* Implement StreamWrapper

* Add local socket UI

* Resolve relative/tilde paths for local socket

* Improve connection error popups & clear state upon reconnecting

TODO: move folder view state to controller like other views.

* Fix playlist queue-all button queueing twice

* Fix queue view resetting scroll position on swaps

Scroll still resets on track removal. This seems hard to fix.

The biggest change of this commit is actually the whole queue update
logic, which is now local-first, server-later. We stay in sync with the
server by carefully monitoring queue version changes and tracking the
expected server-side version as changes are applied locally.

This new logic is much more efficient both for the server and the
client. It also allows us to at least get rid of a near-full refresh of
the queue upon song swaps and removals. Also thanks to song swaps being
performed locally first (and not having to be locally realised later via
a full queue refresh) we have gotten rid of the jump-to-beginning bug
at least for swaps.

For some reason removals will still reset the queue location. Not even a
manual scroll position restoration in g_idle_add_once() helped.

* Work around queue scroll position bug

* Refactor folder view state into library controller

Finishing what the background client refactor set out to do.

* Bump version and update metadata

* Fix Playlist Editor scroll issue too

Same issue as Queue View